Finds the largest number in the list.

+++ src/largest.rs
@@ -1,4 +1,4 @@
-pub fn largest(numbers: &Vec<usize>) -> usize {
+pub fn largest(numbers: &Vec<u64>) -> u64 {
     let mut largest = numbers[0];
 
     for &item in numbers.iter() {
